Improve freeze-dry cleanup algorithm
- moved from using plain skip to a manual count
- until the $exists issue is sorted, to keep the 100 most recent non-criteria-matching freeze-dries, skip may skip ones we want

+++ src/page-storage/cleanup.js
@@ -4,7 +4,7 @@ import { pageDocsSelector } from './'
 
 const freezeDryAttachmentId = 'frozen-page.html'
 
-const getQuery = ({ skipBookmarkPages = false, skip = 100, customFields = [] }) => {
+const getQuery = (skipBookmarkPages = false, customFields = []) => {
     // TODO: Replace this with actual bookmarks flag on page docs from bookmark creation event
     const bookmarksFlagSelector = skipBookmarkPages ? { isBookmarkPage: { $ne: true } } : {}
 
@@ -12,10 +12,9 @@ const getQuery = ({ skipBookmarkPages = false, skip = 100, customFields = [] }) 
         selector: {
             ...pageDocsSelector,
             keepFreezeDry: { $ne: true },
+            isStub: { $ne: true },
             ...bookmarksFlagSelector,
         },
-        skip,
-        sort: [{ _id: 'desc' }],
         fields: ['_id', '_rev', ...customFields],
     }
 }
@@ -25,17 +24,21 @@ const getQuery = ({ skipBookmarkPages = false, skip = 100, customFields = [] }) 
  *  - not in the 100 latest page docs
  *  - not manually specified by user to keep (`keepFreezeDry` flag)
  *  - not associated with created bookmark doc (if user-option set)
+ * @param {number} [keepLimit=100] The number of recent page freeze-dries to keep
  */
-export default async function cleanupFreezeDry() {
+const cleanupFreezeDry = (keepLimit = 100) => async () => {
     const skipBookmarkPages = (await browser.storage.local.get(FREEZE_DRY_BOOKMARKS_KEY))[FREEZE_DRY_BOOKMARKS_KEY]
 
-    const { docs } = await db.find(getQuery({ skipBookmarkPages, customFields: ['_attachments'] }))
+    const { docs } = await db.find(getQuery(skipBookmarkPages, ['_attachments']))
+    let numProcessed = 0
 
-    // Remove freeze dry attachment from docs if present
-    for (const doc of docs) {
+    // Remove freeze dry attachment from docs if present (start from latest)
+    for (const doc of docs.reverse()) {
         // TODO: replace this check with $exists in query
-        if (doc._attachments && doc._attachments[freezeDryAttachmentId]) {
+        if (doc._attachments && doc._attachments[freezeDryAttachmentId] && numProcessed++ > keepLimit) {
             await db.removeAttachment(doc._id, freezeDryAttachmentId, doc._rev)
         }
     }
 }
+
+export default cleanupFreezeDry
